Move views into a separate file to disambiguate the global vars
